Generalize support for self-closing tags.
According to the HTML spec[^1] adding a "/" right before the closing ">"
marks the (foreign) tag as self-closing. Previously the HTML parser
would generate endless nesting with sequences of self-closing tags,
which is particularly annoying with for example Vue templates. But also
beyond that, it made the syntax trees difficult to consume for anything
beyond syntax highlighting when self-closing tags were used in the
input text.

[^1]: https://html.spec.whatwg.org/multipage/syntax.html#start-tags

Bug: https://crbug.com/1383451